Added a fastForward migration method
This is to enable the migrations to be marked as complete without the migrations actually being run – which eg. is something you likely want to do if you install an instance of a website from scratch and already have an up to date schema from the installation.

This new option enables the installation script to mark all migrations as already complete by doing `knex.migrate.fastForward()` to run through all migrations without actually applying any of the changes.

This same pattern could probably be used in other cases as well where a change has already happened outside the migration scripts for one reason or another.

I myself am only interested in pairing it with an installation scripts that sets up a fresh already up to date schema from scratch.