Include lib files
These files were generated by the changes from the previous commit.

There are a lot of files I have not changed in the source that were
included in this build. This probably has happened because I’m using
the latest version from Babel 5.8.23 (babel-core 5.8.25) based in the
rule babel@^5.2.17 from package.json.