MySQL hasTable and hasColumn should work when no database is selected
The form of `show tables` and `show columns` being used required that
the connection specify a database. further, it did not honor a schema
specified via `withSchema`. Calling `hasTable` or `hasColumn` would
result in an error because no database was selected.

By querying information_schema tables for these operations, all of
the following scenarios are supported:
* connection settings specifies database
  * knex.hasTable('table')
  * knex.withSchema('schema').hasTable('table')
  * knex.hasTable('schema.table')
* connection settings does not specify database
  * knex.withSchema('schema').hasTable('table')
  * knex.hasTable('schema.table')

The following scenarios will always return false:
* knex.withSchema('schema').hasTable('schema.table')
* connection settings does not specify a database
  * knex.hasTable('table')