Added requirement of integration tests
There has been many PRs which are changing what kind of queries are sent to DB without any real tests that also checks that sent queries are doing the things that they are supposed to do. To be able to make sure that those new queries actually work, in addition to unit tests it is crucial to have some integration tests as well to be able to guarantee that everything won't be broken after the changes.