use acquireRawConnection and _query to cancel query on timeout
KILL commands have to be sent on a seprate connection, and transaction
clients override acquireConnection to always return the same connection.

Furthermore, it is plausible that some applications may never be able to
acquire a pooled connection for cancellation becuase a query which is
exceeding the timeout may coincidentally exhaust the connection pool.

For example, if a DDL operation is blocking writes, a connection pool
may be quickly exhaused by write operations.

fixes #2211