MSSQL: Transaction Bug
Fixed an issue where rollback() would attempt to rollback a transaction that has not been opened yet.  Calling beginTransaction() will not open the transaction until the first query, and if the first query has not executed yet, rollback() will attempt to roll back a transaction that is not open.

mssql: do not issue rollback if connection is not in transaction