Work-around for nyc interfering w/ Liftoff's ".ts" handling ...
... Specifically: the default configuration for nyc causes it
to add ".ts" to the the `require.extensions` list.

The problem: Liftoff consults `require.extensions` when deciding
which modules need to be preloaded.  Since ".ts" is already on
the list, Liftoff decides that "ts-node" must have already been
preloaded.  Therefore, it hands control off to the Knex CLI
without ever registering Typescript support.  Consequently, the
Knex CLI will explode if it attempts to open a Knexfile that was
written in Typescript.

In other words: this error only occurs when `nyc` (or another
tool) modifies `require.extensions` before handing control over
to the Knex CLI.