change .map to .forEach
It doesn't make sense to use .map here; this allocates and produces an array of `undefined` (the return value of `console.log`) only to be tossed to the garbage collector. Better to use `forEach` if no mutations are applied to the array and the intent is to produce a side effect (logging).