Adding constants for driver names
When testing against the driverName there seems to be a number of
locations where the value being checked against is not the actual
driverName. This may result in tests passing or failing incorrectly.

Adding a constant that has all current driverNames defined so that
matching can be done confidently and with tooling support. It's
effectively an enum.

Updating the tests to use the db-helpers, and having the db-helpers
utilize the drivers for correct naming should help ensure all tests fire
as expected and allow easier refactoring.