Moving connection logic to postgres strategy
Using the newer syntax in the default postgres strategy and inheriting
it into the pg-native.

Also extracting connection creation in postgres dialect

The `pg` library, both standard and `native` clients, offers the option
to return a promise instead of callback when calling `connect`.

The unit tests for `pg` were mocking out the `connect` call, but that
relied on the deprecated call signature in `pg` standard client to work.

The `_acquireOnlyConnection` call will now acquire the connection and
connect to it without any of the additional configuration of the rest of
`acquireRawConnection` to keep the encapsulation clean and make it
easily mockable if necessary.