Improve connection string handling for SQLite3 and Postgresql.
Allow use null as default for sqlite3 to be specified in a connection string
by including it in the query paramters.  An example is:

"sqlite3:///:memory:?useNullAsDefault=true"

Additionally improve PostgreSQL connection string handling by allowing
the searchPath to be specified in the connection string.  An example is:

"postgresql:///?searchPath=pg_temp"

Without these changes it is impossible to allow users to specify these options
via connection strings.