Added getDialectByNameOrAlias and converted repo to use it.
Previous this repo has imported Dialect objects via dynamically generating
a string to require. This does not work well with compilers like ncc. To
solve this and make for a better development experience we add the function
`getDialectByNameOrAlias` to safely import Dialects while making requirement
paths static for better compiler and bundler experiences.