Use correct paths in package.json browser field
The seeder entry is the only one really required to easily build knex for
the browser without shimming `fs`, but I updated the migrations path to
match.

However, there's also `lib/migrations/seed/seed-stub.js`, which actually
seems to provide errors when attempting to use the seed functionality in
a browser, but isn't actually used anywhere

I'm not exactly sure which way is intended.