uuidToBin and binToUuid are switched (#434)
The code examples of uuidToBin and binToUuid where in the wrong order. I switched the heading and description of both so the createTable and  select statements are still correct.
I also changed 'convert a uuid (char(16))' to 'convert a string uuid (char(36))' in uuidToBin to match the description of binToUuid and the corresponding code example.
In addition I changed the specified language of the binToUuid code example from ts to js to match the language definitions of the other code examples.