Fix FOR UPDATE OF with explicit schema (#5053)
When the _tableNames helper was written there were multiple methods
hitting it, but they've since been merged into the single _lockingClause
method. Given that the only really interesting thing it was doing was
prepending the schema name--which is the incorrect behavior this commit
fixes--it didn't seem worth keeping at this point.

I'm not sure you'd say `.withSchema('public')` in a real query but it
lets me exercise the behavior in question without having to learn enough
about knex's test environment to stand up a discrete schema. I suppose
it's plausible that a reasonable maintainer would go "hm that's silly
and redundant" and inadvertently remove coverage for this bug, though.