fix: actually run the CLI tests
Background: #5138 bumped Jake, and as part of the migration moved to
using child_process instead of jake.createExec. However, the events
emitted by Jake and Node when a process exits are different, so the
event handler intended to resolve the promise when the exec finished
never ran.

Since no handles held the process open, it just exited with 0, doing
nothing.

I chose to utilize the callback form of `child_process.exec` since
it already does the work these functions were doing (primarily,
accumulating the contents of stdout and stderr).

With this change, `npm test:cli` runs the tests again.