fix: return consistently-shaped data from Migrator.js
Background: #5541 introduced lifecycle hooks into the migrator, but
avoids running them if there are no migrations to execute. It returns
an empty array from `_runBatch` in this case, but if there is work
to do it returns the result of `this._waterfallBatch`, which differs:
`_waterfallBatch` returns `[0, []]` when it has nothing to do.

The return value here is consumed in the CLI by `migrate:up`,
`migrate:rollback`, and `migrate:down`, throwing an undefined
property access error on the missing second argument.

I chose to just initialize the return value to `[0, []]` instead of
`[]` because I'm uncertain whether anything else depends on this
behavior. Plausibly, users accessing migrations via an exposed
API may have similar issues.