fix: ensure openKnexfile always throws an instance of Error
The only possible way I can find to reproduce #6082 is to throw
`Object.create(null)` from a knexfile. I'm fairly confident that
users arent doing something like that directly, but I could
conceive of it happening as a consequence of a library they
used or imported. The other code paths appear to get wrapped
in Error.

This commit just ensures that if anything is thrown from
the code that's out of our control (knexfile), it is always
thrown as an Error instance. This allows the common `.catch(exit)`
pattern in the rest of the file to convey useful information
to the end-user in this edge case.