fix: protect cli-config-utils's `exit` against unexpected arguments
Fixes #6082.

The exit() function expects either an Error or a string, and if it
receives something that is neither it can misbehave and obscure
the real reason it was called.

If we receive an error, we now also check for and log the contents
of `.nonErrorValue`, added by the previous commit in the case of
errors caught from the knexfile. This ensures that we present as much
information as we can to the user in such a case.

In the non-error case, we check that the argument is a string before
implicitly treating it as one, and then do as we did before.

In the final case (not an Error, not a string), we log a new message
indicating that it's an unexpected case. Here, we log the value
directly (without coloring), since `console.log` should be infallible
but `color.red` is not. `console.log` will also generally present
arbitrary data in a more useful-to-read way in Node and in browsers
(for example: `[Object: null prototype] { foo: 'bar' }` instead of
just `{ foo: 'bar' }`)