General: Drop extra integration test tables in cleanup
In Oracle, dropping the "accounts" table was failing because it was
referenced by the "test_foreign_table_two" table (which is created
in the schema/misc tests). Ensure it is dropped so that accounts
can be dropped as expected.

I also identified "check_test" as another table being created by
raw tests (and not the tableCreatorHelper code) and added it to
the list of tables to clean up.