MySQL: ensure query contents are unique for timeout tests
The last timeout test, which abandons the connection, was leaving
the sleep query running. Since both "mysql" and "mysql2" drivers
are connecting to the same database for integration testing, when
both were executed in the same test run, some "mysql2" tests were
seeing these hanging queries and failing, when they were actually
started by a different test against a different database client.

I added comments to the queries to ensure that they are unique
per-test and per-driver, so that the timeout tests are only
asserting against their own queries.