Use a proper port number for connection string parse assertions
`parse-connection.js`'s `tryParse` uses `new URL` to parse URLs.
Even though the "port" property of a URL instance is a string,
the string is of a non-zero-prefixed integer. These tests were
failing since they asserted that "0000" parsed to "0000", but
in actuality they parsed to "0". Rather than adjust the test
to expect "0", I changed the port to something that would
result in the same thing as the input; that way, we're not
testing the behavior of the URL constructor's parsing, but
just that what _we_ care about is true (the port supplied
in the connection string becomes the value in the object)