Fix assertions for non-enumerable password property of connection config
https://github.com/knex/knex/pull/5559 adjusted makeKnex to store the
`password` property as non-enumerable. As a consequence, tape's `t.deepEqual`
was failing when comparing the rendered config with the expectation.

In the "parse connection string" test, I added an extra expectation to
ensure the password was indeed correctly extracted from the connection string.

In the other tests, I just removed the password property from the expectation,
since they aren't testing the password specifically and we're already covered
on that front.