Refactor SQLite to allow setting the foreign_keys pragma for a transaction
The SQLite schema builder uses a pretty standard workaround to drop columns:
- Create a copy of the target table without the column in question
- Copy over the data
- Drop the old table and rename the new one

In order to prevent rows in other tables with ON CASCADE DELETE from
being dropped when the old table gets dropped (sqlite executes an
implicit DELETE on all rows when dropping a table), the DDL helper
attempted to disable foreign key enforcement before performing
this dance, and enable it after.

The DDL helper performs this work in a transaction, which is reasonable,
except that "PRAGMA foreign_keys" has no effect in a transaction:
https://sqlite.org/pragma.html#pragma_foreign_keys

In order to ensure that we can "have a transaction with foreign key
enforcement disabled", I extended the Transaction_Sqlite class to
accept a new config option, `enforceForeignCheck`. When set to
a boolean, the transaction that the callback receives will have
its `foreign_keys` pragma set to ON (for true) or OFF (for false).
If not defined, no change is made.

In order to prevent future error, I also introduced a private
"_strict()" method on the Client_SQLite3 class, which idempotently
returns a version of the client which will refuse to create a
transaction unless the `enforceForeignCheck` property is specified.
This should guard against people making changes or adding new code
but not knowing they need to use this mechanism. The `.ddl()` method
returns a "strict-mode" client, and the two uses of transactions
in the SQLite3_DDL class were updated to supply the required option.

I had to refactor the base Transaction class a little so that
the Transaction_Sqlite class could override and wrap the business
logic with setup and cleanup code.

When in strict mode, Transaction_Sqlite will also throw an error
if it attempts to create a nested transaction _and also_ needs
to change the foreign_keys pragma. This may cause knex to start
failing for some people who have been using it in a way that
might have caused them data loss.

Users who start hitting this error are advised to run their
migrations with `disableTransactions: true`; the migration
runner doesn't have any way to know if the code that it runs
will be dropping columns using this workaround. They should
then also update their migrations to use transactions manually
where needed.

Future work may be able to improve this, but I want to keep
the code change here as minimal as possible for now.

All changed code has 100% coverage.

Fixes #6213