Use cross-env to set env vars
This would be improved by cli-testlab supporting environment
variables, but for now this is the easiest fix. Note that
this only works when the tests are run via npm (which is
how the `cross-env` binary comes to exist in the path)

This seems reasonable for now, since that's the most common
way that tests _will_ be run, and even if executing `mocha`
manually, it'd be via e.g. `npx mocha`