Fix `test:cli` tests (#6310)
* Run all CLI tests; prevent color output from failing tests

- Extracts the mocha-based CLI tests to a separate file that
  can be readily run on its own
- Adds the mocha-based CLI tests to `npm run test:cli`
- Updates a couple CLI tests that were comparing colored output
  to run with `NO_COLOR=true` and expect uncolored output
  These tests were previously failing out of the box when
  running `npm run test:cli`

* Remove unused reference to `colorette`

* Use cross-env to set env vars

This would be improved by cli-testlab supporting environment
variables, but for now this is the easiest fix. Note that
this only works when the tests are run via npm (which is
how the `cross-env` binary comes to exist in the path)

This seems reasonable for now, since that's the most common
way that tests _will_ be run, and even if executing `mocha`
manually, it'd be via e.g. `npx mocha`