Fix various failing tests (#6313)
* Update test to match both old and new error descriptions

I'm assuming this test used to pass, but the error message I get from
the version of oracle using the documented setup includes "or closed".
Change to a regex that matches either.

* MSSQL: Drop "test" table before each run

For some reason, the "afterEach" hook is not enough to leave the
database clean of this table which is used in the "column default
handling" tests. Ensure it's removed before running the tests.

* General: Drop extra integration test tables in cleanup

In Oracle, dropping the "accounts" table was failing because it was
referenced by the "test_foreign_table_two" table (which is created
in the schema/misc tests). Ensure it is dropped so that accounts
can be dropped as expected.

I also identified "check_test" as another table being created by
raw tests (and not the tableCreatorHelper code) and added it to
the list of tables to clean up.

* MySQL: ensure query contents are unique for timeout tests

The last timeout test, which abandons the connection, was leaving
the sleep query running. Since both "mysql" and "mysql2" drivers
are connecting to the same database for integration testing, when
both were executed in the same test run, some "mysql2" tests were
seeing these hanging queries and failing, when they were actually
started by a different test against a different database client.

I added comments to the queries to ensure that they are unique
per-test and per-driver, so that the timeout tests are only
asserting against their own queries.

* Use a proper port number for connection string parse assertions

`parse-connection.js`'s `tryParse` uses `new URL` to parse URLs.
Even though the "port" property of a URL instance is a string,
the string is of a non-zero-prefixed integer. These tests were
failing since they asserted that "0000" parsed to "0000", but
in actuality they parsed to "0". Rather than adjust the test
to expect "0", I changed the port to something that would
result in the same thing as the input; that way, we're not
testing the behavior of the URL constructor's parsing, but
just that what _we_ care about is true (the port supplied
in the connection string becomes the value in the object)

* Fix assertions for non-enumerable password property of connection config

https://github.com/knex/knex/pull/5559 adjusted makeKnex to store the
`password` property as non-enumerable. As a consequence, tape's `t.deepEqual`
was failing when comparing the rendered config with the expectation.

In the "parse connection string" test, I added an extra expectation to
ensure the password was indeed correctly extracted from the connection string.

In the other tests, I just removed the password property from the expectation,
since they aren't testing the password specifically and we're already covered
on that front.

* Exclude `better-sqlite3` from the db-setup tests too

`sqlite3` was already excluded, because it succeeds always. I assume
`better-sqlite3` was added later but not excluded from this test.