Update type test to correctly fail on a regression
`tsd` does not fail assertions when a type resolves to `any`, despite
the claim in https://github.com/tsdjs/tsd/issues/82

I tried on latest and on the current version, and since it didn't
solve the problem, I kept the version where it is. Separately,
we should migrate to `expect-type`.

`vitest` uses `expect-type`, so it is heavily used and well-maintained,
and `expect-type` provides a more robust API. I hope to migrate our
tests to `vitest` down the line when we're ready to make broad code
changes.