Move __knexUid and __knexTxId away from connection instances
In general, mutating "someone else's" data is undesirable.

These properties appear to be used for log correlation and no
actual functionality, so this change should be pretty safe.

Instead of mutating connection instances with properties,
associate them indirectly via a weakmap and expose a singleton
to allow retrieval. Motivated by #5288, this also fixes the
"__knexTxId persists after the transaction is complete" behavior.