Add sqlite to the exclusion list
While sqlite3 _does_ have support for json query functions, it does
not actually recognize "json" as a column type in e.g. create table
(see https://sqlite.org/datatype3.html#determination_of_column_affinity)

... and will, in fact, accept _anything_ as the typename of a column.
unknown columns just fall back to the "NUMERIC" affinity, which is
the wildcard that accepts anything, including strings:

https://sqlite.org/datatype3.html#type_affinity
> A column with NUMERIC affinity may contain values using all five storage
> classes

In sum, like dates, there's nothing that a client can really do to
"know" whether data should be serialized or deserialized as json;
sqlite relies on the user to know what they put there and what to
do with it when they get it back. This showed up in the test as
a special case for sqlite where we manually serialize and deserialize
the value as json - thus defeating the thing being tested...