Clean up "order by null" tests
We have unit tests on the sql syntax already, and the whole
`.testSql(...)` flow is vulnerable to leaving out dialects.

Instead, simply assert the behavior of the returned values
directly. Cover all combinations of nulls first/last,
column ascending/descending, on all dialects.

Use the column created for the purpose to determine
the order of the returned results, rather than `id`
which is implicit and not an expected serial integer
from 0 on cockroachdb.

Explicitly specify the order of the comparison column
to ensure stable ordering from all databases rather
than leaving the behavior undefined