Fix mssql "order by nulls"
The order produced by the code was backwards. Simplify the code
to be more explicit about the null-ordering by using the same
expression in both cases and just changing 'asc' or 'desc' on
the expression's result. This avoids reliance on database
default behavior and/or us knowing what that even is.