Remove `update_gitignore_for_tsc_output` script
This script's original intent was to manually add typescript build
artifacts to the `lib/.gitignore` file. However, the decision was
later made (commit ca3dd19f4f87b8e3646f9b190ed2738742133949) to
check these files in. This script was never removed, so when a
developer runs `npm build` it is adding .js files that we intend
to check in to lib/.gitignore

Since the rest of .gitignore is satisfiable with just wildcard
rules, I've removed the script entirely.